+++ src/commands/view.ts
@@ -2,7 +2,7 @@ import type { Command } from 'commander';
 import { BrowserBehavior, startServer } from '../server/server';
 import telemetry from '../telemetry';
 import { setupEnv } from '../util';
-import { setConfigDirectoryPath } from '../util/config';
+import { setConfigDirectoryPath } from '../util/config/manage';
 
 export function viewCommand(program: Command) {
   program
