+++ src/util/config/manage.ts
@@ -2,9 +2,9 @@ import * as fs from 'fs';
 import yaml from 'js-yaml';
 import * as os from 'os';
 import * as path from 'path';
-import { getEnvString } from '../envars';
-import type { UnifiedConfig } from '../types';
-import { orderKeys } from './json';
+import { getEnvString } from '../../envars';
+import type { UnifiedConfig } from '../../types';
+import { orderKeys } from '../json';
 
 let configDirectoryPath: string | undefined = getEnvString('PROMPTFOO_CONFIG_DIR');
