+++ test/redteam/commands/generate.test.ts
@@ -4,12 +4,12 @@ import logger from '../../../src/logger';
 import { synthesize } from '../../../src/redteam';
 import { doGenerateRedteam } from '../../../src/redteam/commands/generate';
 import type { RedteamCliGenerateOptions } from '../../../src/redteam/types';
-import { writePromptfooConfig } from '../../../src/util/config';
+import { writePromptfooConfig } from '../../../src/util/config/manage';
 
 jest.mock('fs');
 jest.mock('yaml');
 jest.mock('../../../src/redteam');
-jest.mock('../../../src/util/config');
+jest.mock('../../../src/util/config/manage');
 jest.mock('../../../src/logger');
 jest.mock('../../../src/telemetry');
